Refactoring; fix all warnings
 - Add missing `const` qualifiers

 - Rename variables to prevent shadowing

 - Use Race in MapBorderRenderer

 - Reintroduce `static` keyword to global constants

 - Combine `window.use()` with ctor

 - Remove implicit casts

 - Remove unused variables

 - Disable warnings from third-party libraries

 - Add named constants for map border texture indices

+++ Open-Rival/src/BinaryFileReader.cpp
@@ -17,7 +17,7 @@ namespace Rival {
     }
 
     std::uint8_t BinaryFileReader::readByte() {
-        return in.get();
+        return static_cast<std::uint8_t>(in.get());
     }
 
     std::uint16_t BinaryFileReader::readShort() {
